add a few new libaries to support icons in buttons
we added react-icons for our icons and then added those icons to the buttons but it caused some issues. the 'classnames' library wasn't behaving as expected when we were trying to override the 'text-white' class on our buttons. we fixed this by installing the 'tailwind-merge' library and wrapping the classnames function